factor out apikeygate into own file
